[Rename] restaurant 패키지명 수정
