+++ src/main/java/com/allclear/tastytrack/domain/restaurant/repository/RawRestaurantRepository.java
@@ -1,6 +1,6 @@
-package com.allclear.tastytrack.domain.restraunt.repository;
+package com.allclear.tastytrack.domain.restaurant.repository;
 
-import com.allclear.tastytrack.domain.restraunt.entity.RawRestaurant;
+import com.allclear.tastytrack.domain.restaurant.entity.RawRestaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RawRestaurantRepository extends JpaRepository<RawRestaurant, Long> {
